Allow deployment of cluster with security hardening configuration (#90)
* Allow toogle between secure ports and unsecure

Change-Id: I3aae090964f16951b94e8fabb7c38ef9696fee0d

* Expose openSSL config

Change-Id: If67c83e1060b73e2a1feaf4c59e233d0360b7970

* Expose ability to secure internal communication

Add the ability to configure the cluster so that internal communication
is encrypted and password-protected.

Change-Id: If40cf4fe4075f870e725c73e7af7930f3c987dc9

* Add ability to set ZooKeeper ACL config

ClickHouse has the ability to set ZooKeeper access control to the nodes
it is using. This provides additional layer of security since it makes
sure that the zookeeper nodes used by clickhouse is password protected
and cannot be modified by anyone else.

Change-Id: I28cf5ea07d1264befb3628fb38a372c7ce2b8552

* Fix connection on secure tcp port

If user configures secure tcp port it would be best to check connection
using the secure port. Also communicate with the database on the secure
connection to make sure it works.

Change-Id: I8efe34b5279d63438a290e9a3919d5afbef01800

* Add security hardening documentation to README.md

Add security hardening documentation. Provides the ability to enforce
better security practices for network encryption and secret/credential
based access for zookeeper, data replication and distributed queries.

Change-Id: I5dc416ee5c48123af27cf405adf7a877aa2e6204